+++ vyuh_cli/lib/src/commands/create/commands/feature_sanity_schema/index.dart
@@ -1,2 +1,0 @@
-export 'feature_sanity_schema_command.dart';
-export 'feature_sanity_schema_subcommand.dart';