+++ vyuh_cli/lib/src/commands/create/commands/index.dart
@@ -1,3 +1,3 @@
 export 'feature/index.dart';
-export 'feature_sanity_schema/index.dart';
 export 'project/index.dart';
+export 'schema/index.dart';