+++ vyuh_cli/lib/src/commands/create/templates/index.dart
@@ -1,4 +1,4 @@
 export 'feature/index.dart';
-export 'feature_sanity_schema/index.dart';
 export 'project/index.dart';
+export 'schema/index.dart';
 export 'template.dart';