+++ vyuh_cli/lib/src/commands/create/templates/schema/index.dart
@@ -1,2 +1,2 @@
-export 'feature_sanity_schema_template.dart';
+export 'schema_template.dart';
 export 'vyuh_feature_sanity_schema_bundle.dart';