+++ vyuh_cli/CHANGELOG.md
@@ -3,4 +3,4 @@
 - First release with a set of core commands
   - `vyuh create project`
   - `vyuh create feature`
-  - `vyuh create feature-sanity-schema`
+  - `vyuh create schema`